fix: rank:ndcg のラベルを整数にキャスト (#61)
* feat: コース適性の走り方推論特徴量5個を追加

プロ予想家的な「走り方からのコース適性推論」を特徴量化。
既存のcalc_course_aptitude_featuresに以下5個を追加:

- f_course_type_rate: 同じcourse_type(tight/spacious等)での好走率
- f_tight_vs_spacious: 小回りvs大箱の成績差（正=小回り得意）
- f_uphill_severity_rate: 同じ坂レベル(steep/medium/none)での好走率
- f_corner_position_change: 小回りでの3→4角位置変動（器用さ指標）
- f_agari_rank_at_type: 同タイプコースでの上がり相対順位

次回XGB再学習(build_training_data→train_xgb)で自動取込される。


* fix: rank:ndcg のラベルを整数にキャスト（XGBoost新版対応）

XGBoostの新バージョンではrank:ndcgのラベルが0以上の整数でなければ
エラーになる。placeがfloat型で渡されていたためint変換を追加。


---------

+++ src/tools/train_ranking_model.py
@@ -59,7 +59,7 @@ def build_ranking_training_data(df, feat_cols):
             continue
 
         field_size = n
-        scores = (field_size - valid['place'] + 1)  # 整数ラベル [1, field_size]（ndcg評価指標の要件）
+        scores = (field_size - valid['place'].astype(int) + 1).astype(int)  # 非負整数 [1, field_size]
 
         X_list.append(valid[feat_cols].fillna(5.0).values)
         y_list.append(scores.values)
